requirements.txt
asgiref==3.2.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django==3.1.1
django-allauth==0.50.0
django-filter==21.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.3
itypes==1.2.0
Jinja2==3.1.1
Markdown==3.3.6
MarkupSafe==2.1.1
oauthlib==3.2.0
openapi-codec==1.3.2
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.0
